mdtests macro wip
