+++ Cargo.toml
@@ -4,33 +4,34 @@
 members = [
     "benches",
     "datastores/gossip_kv",
-    "hydro_deploy/core",
-    "hydro_deploy/hydro_cli",
-    "hydro_deploy/hydro_cli_examples",
-    "hydro_deploy/hydroflow_deploy_integration",
-    "dfir_rs",
-    "dfir_datalog",
     "dfir_datalog_core",
+    "dfir_datalog",
     "dfir_lang",
     "dfir_macro",
+    "dfir_rs",
+    "hydro_deploy/core",
+    "hydro_deploy/hydro_cli_examples",
+    "hydro_deploy/hydro_cli",
+    "hydro_deploy/hydroflow_deploy_integration",
     "hydro_lang",
     "hydro_std",
-    "hydro_test",
-    "hydro_test_local",
     "hydro_test_local_macro",
-    "lattices",
+    "hydro_test_local",
+    "hydro_test",
+    "include_mdtests",
     "lattices_macro",
+    "lattices",
     "multiplatform_test",
     "pusherator",
     "relalg",
-    "stageleft",
     "stageleft_macro",
-    "stageleft_test",
     "stageleft_test_macro",
+    "stageleft_test",
     "stageleft_tool",
+    "stageleft",
     "topolotree",
-    "variadics",
     "variadics_macro",
+    "variadics",
     "website_playground",
 ]
 exclude = ["template"]
