+++ dfir_rs/Cargo.toml
@@ -33,11 +33,11 @@ required-features = [ "python" ]
 bincode = "1.3.1"
 byteorder = "1.3.2"
 bytes = "1.1.0"
-futures = "0.3.0"
-hydroflow_deploy_integration = { optional = true, path = "../hydro_deploy/hydroflow_deploy_integration", version = "^0.12.0" }
 dfir_datalog = { optional = true, path = "../dfir_datalog", version = "^0.12.0" }
 dfir_lang = { path = "../dfir_lang", version = "^0.12.0", optional = true }
 dfir_macro = { optional = true, path = "../dfir_macro", version = "^0.12.0" }
+futures = "0.3.0"
+hydroflow_deploy_integration = { optional = true, path = "../hydro_deploy/hydroflow_deploy_integration", version = "^0.12.0" }
 itertools = "0.10.0"
 lattices = { path = "../lattices", version = "^0.6.0", features = [ "serde" ] }
 pusherator = { path = "../pusherator", version = "^0.0.11" }
@@ -78,9 +78,9 @@ dfir_lang = { path = "../dfir_lang", version = "^0.12.0", features = [ "clap-der
 chrono = { version = "0.4.20", features = [ "serde", "clock" ], default-features = false }
 clap = { version = "4.5.4", features = [ "derive" ] }
 colored = "2.0"
+include_mdtests = { path = "../include_mdtests", version = "^0.0.0" }
 insta = "1.39"
 multiplatform_test = { path = "../multiplatform_test", version = "^0.5.0" }
-wasm-bindgen-test = "0.3.33"
 rand = { version = "0.8.0", features = [ "small_rng" ] }
 rand_distr = "0.4.3"
 regex = "1.10.4"
@@ -88,6 +88,7 @@ static_assertions = "1.0.0"
 time = "0.3"
 tracing-subscriber = { version = "0.3", features = [ "env-filter" ] }
 trybuild = "1"
+wasm-bindgen-test = "0.3.33"
 zipf = "7"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
