+++ hydro_test/Cargo.toml
@@ -25,14 +25,12 @@ stats-ci = "0.1.1"
 stageleft_tool = { path = "../stageleft_tool", version = "^0.6.0" }
 
 [dev-dependencies]
-ctor = "0.2"
-insta = "1.39"
-hydro_deploy = { path = "../hydro_deploy/core", version = "^0.12.0" }
-hydro_lang = { path = "../hydro_lang", version = "^0.12.0", features = [
-    "deploy",
-] }
-futures = "0.3.0"
 async-ssh2-lite = { version = "0.5.0", features = ["vendored-openssl"] }
-dfir_macro = { path = "../dfir_macro", version = "^0.12.0" }
-tokio-test = "0.4.4"
+ctor = "0.2"
 dfir_rs = { path = "../dfir_rs", version = "^0.12.0" }
+futures = "0.3.0"
+hydro_deploy = { path = "../hydro_deploy/core", version = "^0.12.0" }
+hydro_lang = { path = "../hydro_lang", version = "^0.12.0", features = [ "deploy" ] }
+include_mdtests = { path = "../include_mdtests", version = "^0.0.0" }
+insta = "1.39"
+tokio-test = "0.4.4"
