+++ include_mdtests/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "include_mdtests"
+publish = true
+version = "0.0.0"
+documentation = "https://docs.rs/include_mdtests/"
+description = "Include markdown for testing, via doctests."
+edition = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[lints]
+workspace = true
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2.0.46", features = [ "parsing" ] }
+proc-macro2 = "1.0.74"
+quote = "1.0.35"
+glob = "0.3.2"
